break up test_registration_detail two smaller files ~500 lines
